// frontend/src/types/cars.types.ts
export interface CarEntity {
  id: number;
  year: number;
  numberOfSeats: number;
  driveType: string;
  fuelType: string;
  dailyPrice: number;
  power: number;
  brand: string;
  model: string;
  engineCapacity: string;
  color: string;
  transmission: string;
  fuelUsageCity: string;
  fuelUsageOutcity: string;
  imageUrl: string;
  index: number;
  localisation: string;
}